# Perf work on interger array