--- a/src/realm/array_flex.cpp
+++ b/src/realm/array_flex.cpp
@@ -47,14 +47,15 @@ void ArrayFlex::copy_data(const Array& arr, const std::vector<int64_t>& values,
     REALM_ASSERT_DEBUG(arr.m_encoder.get_encoding() == Encoding::Flex);
 
     auto h = arr.get_header();
-    auto v_width = NodeHeader::get_elementA_size<Encoding::Flex>(h);
-    auto ndx_width = NodeHeader::get_elementB_size<Encoding::Flex>(h);
-    auto v_size = values.size();
-    auto data = (uint64_t*)arr.m_data;
-    auto offset = static_cast<size_t>(v_size * v_width);
+    const auto& encoder = arr.get_encoder();
+    const auto v_width = encoder.m_v_width;
+    const auto ndx_width = encoder.m_ndx_width;
+    const auto v_size = values.size();
+    const auto data = (uint64_t*)arr.m_data;
+    const auto offset = static_cast<size_t>(v_size * v_width);
     bf_iterator it_value{data, 0, v_width, v_width, 0};
     bf_iterator it_index{data, offset, ndx_width, ndx_width, 0};
-    for (size_t i = 0; i < values.size(); ++i) {
+    for (size_t i = 0; i < v_size; ++i) {
         it_value.set_value(values[i]);
         REALM_ASSERT_DEBUG(sign_extend_field(v_width, it_value.get_value()) == values[i]);
         ++it_value;
@@ -88,36 +89,30 @@ int64_t ArrayFlex::get(const Array& arr, size_t ndx) const
 {
     REALM_ASSERT_DEBUG(arr.is_attached());
     REALM_ASSERT_DEBUG(arr.is_encoded());
-    const auto v_width = arr.m_encoder.m_v_width;
-    const auto v_size = arr.m_encoder.m_v_size;
-    const auto ndx_width = arr.m_encoder.m_ndx_width;
-    const auto ndx_size = arr.m_encoder.m_ndx_size;
-    return do_get((uint64_t*)arr.m_data, ndx, v_width, ndx_width, v_size, ndx_size);
+    const auto& encoder = arr.m_encoder;
+    const auto v_width = encoder.m_v_width;
+    const auto v_size = encoder.m_v_size;
+    const auto ndx_width = encoder.m_ndx_width;
+    const auto ndx_size = encoder.m_ndx_size;
+    const auto mask = encoder.width_mask();
+    return get(arr.m_data, ndx, v_width, ndx_width, v_size, ndx_size, mask);
 }
 
-int64_t ArrayFlex::get(const char* data, size_t ndx, size_t v_width, size_t v_size, size_t ndx_width,
-                       size_t ndx_size) const
+int64_t ArrayFlex::get(const char* data, size_t ndx, size_t v_width, size_t v_size, size_t ndx_width, size_t ndx_size,
+                       size_t mask) const
 {
-    return do_get((uint64_t*)data, ndx, v_width, v_size, ndx_width, ndx_size);
-    //    using Encoding = NodeHeader::Encoding;
-    //    REALM_ASSERT_DEBUG(NodeHeader::get_kind(h) == 'B');
-    //    REALM_ASSERT_DEBUG(NodeHeader::get_encoding(h) == Encoding::Flex);
-    //    //size_t v_size, ndx_size, v_width, ndx_width;
-    //    //get_encode_info(h, v_width, ndx_width, v_size, ndx_size);
-    //    const auto data = (uint64_t*)(NodeHeader::get_data_from_header(h));
-    //    return do_get(data, ndx, v_width, ndx_width, v_size, ndx_size);
+    return do_get((uint64_t*)data, ndx, v_width, v_size, ndx_width, ndx_size, mask);
 }
 
 int64_t ArrayFlex::do_get(uint64_t* data, size_t ndx, size_t v_width, size_t ndx_width, size_t v_size,
-                          size_t ndx_size)
+                          size_t ndx_size, size_t mask)
 {
     if (ndx >= ndx_size)
         return realm::not_found;
     const uint64_t offset = v_size * v_width;
     const bf_iterator it_index{data, static_cast<size_t>(offset + (ndx * ndx_width)), ndx_width, ndx_width, 0};
     const bf_iterator it_value{data, static_cast<size_t>(v_width * it_index.get_value()), v_width, v_width, 0};
-    auto v = sign_extend_field(v_width, it_value.get_value());
-    return v;
+    return sign_extend_field_by_mask(mask, it_value.get_value());
 }
 
 void ArrayFlex::get_chunk(const Array& arr, size_t ndx, int64_t res[8]) const
@@ -134,27 +129,4 @@ void ArrayFlex::get_chunk(const Array& arr, size_t ndx, int64_t res[8]) const
     for (; index < 8; ++index) {
         res[index++] = get(arr, i++);
     }
-}
-
-bool inline ArrayFlex::get_encode_info(const char* h, size_t& v_width, size_t& ndx_width, size_t& v_size,
-                                       size_t& ndx_size)
-{
-    using Encoding = NodeHeader::Encoding;
-    REALM_ASSERT_DEBUG(NodeHeader::get_encoding(h) == Encoding::Flex);
-    v_size = NodeHeader::get_arrayA_num_elements<Encoding::Flex>(h);
-    ndx_size = NodeHeader::get_arrayB_num_elements<Encoding::Flex>(h);
-    v_width = NodeHeader::get_elementA_size<Encoding::Flex>(h);
-    ndx_width = NodeHeader::get_elementB_size<Encoding::Flex>(h);
-    return true;
-}
-
-std::vector<int64_t> ArrayFlex::fetch_all_values(const Array& arr) const
-{
-    REALM_ASSERT_DEBUG(arr.is_attached());
-    REALM_ASSERT_DEBUG(arr.m_encoder.get_encoding() == NodeHeader::Encoding::Flex);
-    std::vector<int64_t> values;
-    values.reserve(arr.m_size);
-    for (size_t i = 0; i < arr.m_size; ++i)
-        values.push_back(get(arr, i));
-    return values;
 }