## What, How & Why?
1. Avoid to create the sign mask for every ArrayEncode::get() has made fetching values from encoding arrays as fast as Array::get() for randomized inputs or even faster. TODO: explore 1.N inputs and N.1 too.
2. The main bottleneck appears to be `ArrayEncode::find()` which is ~5/6x slower at the moment. The amount of accessing we do, to the same 64bit word, is what is dominating. We need to consume more values within the same word and make the searching loop faster.
3. `ArrayEncode::get()`  is now as fast as `Array::get()` . 

Performance wise: These 2 tests are still much slower but significantly better than the current implementation. We still have several slow-downs, but not as big. 
It is probably due to the same problem, we need to reduce the number of accesses to the same 64bit word.

```
Req runs:    5  QueryIntEquality (MemOnly, EncryptionOff):                           * min 389.99ms (+279.12%)          max 403.66ms (+269.91%)        * med 397.93ms (+269.29%)        * avg 397.09ms (+270.90%)          stddev  5.12ms (+102.88%)     ==> was 800%
   
Req runs:   14  QueryEqual<timestamp><NonNullable><NonIndexed> (MemOnly, EncryptionOff):   * min  31.54ms (+367.61%)          max  38.53ms (+305.45%)        * med  32.33ms (+343.13%)        * avg  33.04ms (+345.98%)          stddev  1.91ms (+256.81%)     ==> was 800%
``` 


## ☑️ ToDos
* [ ] 📝 Changelog update
* [ ] 🚦 Tests (or not relevant)
* [ ] C-API, if public C++ API changed
* [ ] `bindgen/spec.yml`, if public C++ API changed